【实验2】 results_gpus.txt =====>3060_gpu_2_70_1k_81k_double.txt

+++ def.h
@@ -10,7 +10,7 @@
 #pragma once
 
 //unified data type
-typedef  float __DATA_TYPE;
+typedef  double __DATA_TYPE;
 
 #include "windows.h"
 #include <stdio.h>
